+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Tasks with a status "app not found" are removed after 1h as they provide no value in history
 - Tasks with a status "in progress" are marked as "aborted" after 1h since that's an indicator of argo-watcher crash
+- Show full time instead of relative on history page
 
 ## [0.1.1] - 2022-07-27
 ### Fixed